Quotes integrated with zenquotes API and create vite.config.js for solve CORS https to http
