+++ src/components/Quotes/quote.css
@@ -1,0 +1,22 @@
+.quote-container {
+    margin-top: 10px;
+    padding: 10px;
+    text-align: center;
+    font-style: italic;
+    border-radius: 5px;
+}
+  .quote {
+    font-size: 1.5rem;
+    color:rgb(155, 153, 153);
+  }
+  
+  .author {
+    font-size: 12px;
+    font-weight: bold;
+    margin-top: 5px;
+    color:white;
+  }
+  p{
+    color:white;
+  }
+
