+++ src/main.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
-import TimeComponent from './components/appTime/appTime.jsx'
+import TimeComponent from './components/appTime.jsx'
 
 
 ReactDOM.createRoot(document.getElementById('app')).render(
